+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `Container::restore`, `Container::resize`, `Container::export`
 * Add support for `!=` filters like `label!=key=val` or `label!=key`
 * Add missing filter function `NetworkPruneOptsBuilder::filters`
+* Fix `Pods::stats` endpoint
 
 # 0.5.0
 * Fix some generated models